Support use of a local CORS proxy in demo
