+++ demo/serve/defs.ts
@@ -6,5 +6,6 @@ export namespace IDemo {
     browsingContextId: string;
     shellManager: IShellManager;
     targetDiv: HTMLElement;
+    useLocalCors: boolean;
   }
 }
