+++ demo/serve/index_remote_cors.ts
@@ -1,0 +1,6 @@
+import { runDemo } from './demo';
+import './style/demo.css';
+
+document.addEventListener('DOMContentLoaded', async () => {
+  await runDemo(false);
+});
